requirements.txt
# AI Image Analyzer v2.0 Requirements
# Desktop GUI Application

# Core AI and ML libraries
google-generativeai>=0.8.0  # Gemini API (cloud fallback)
pillow>=10.0.0              # Image processing

# Image Quality Assessment
torch>=2.0.0                # PyTorch for GPU acceleration
torchvision>=0.15.0         # Vision models
pyiqa>=0.1.7                # Image Quality Assessment

# HTTP requests (for Ollama API)
requests>=2.31.0

# Image metadata handling
piexif>=1.1.3               # EXIF metadata manipulation
PyExifTool>=0.5.6           # ExifTool wrapper for advanced metadata

# System utilities
psutil>=5.9.0               # System monitoring and GPU detection
python-dotenv>=1.0.0        # Environment variable management

# GUI framework (included with Python)
# tkinter - Built into Python, no separate installation needed

# Note: Ollama is installed separately - see README installation instructions
